% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{process_bracket_entries}
\alias{process_bracket_entries}
\title{Extract bracket annotations
Moves bracket-style annotations of the form from taxon_clean
to tax_epithet.}
\usage{
process_bracket_entries(df)
}
\arguments{
\item{df}{A data frame with at least a taxon column. The cleaning
schema is ensured internally before processing.}
}
\value{
The input data frame with updated taxon_clean and tax_epithet.
}
\description{
Extract bracket annotations
Moves bracket-style annotations of the form from taxon_clean
to tax_epithet.
}
