% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{process_sp_entries}
\alias{process_sp_entries}
\title{Extract sp./spp. designations
Moves sp., spp., sp1, and similar designations from taxon_clean
to tax_epithet.}
\usage{
process_sp_entries(df)
}
\arguments{
\item{df}{A data frame with at least a taxon column. If taxon_clean,
tax_epithet, or uncertain are missing, they are created internally.}
}
\value{
The input data frame with updated taxon_clean and tax_epithet.
}
\description{
Extract sp./spp. designations
Moves sp., spp., sp1, and similar designations from taxon_clean
to tax_epithet.
}
