% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vernacular.R
\name{vernacular_list}
\alias{vernacular_list}
\title{List all vernacular-to-taxon mappings}
\usage{
vernacular_list()
}
\value{
A named character vector (name = common name, value = taxon).
}
\description{
List all vernacular-to-taxon mappings
}
