% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%&\%}
\alias{\%&\%}
\title{String concatenation operator}
\usage{
a \%&\% b
}
\arguments{
\item{a}{A character string.}

\item{b}{A character string.}
}
\value{
A character string: \code{paste0(a, b)}.
}
\description{
Shorthand for \code{paste0(a, b)}.
}
