% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{move_morphological_descriptors}
\alias{move_morphological_descriptors}
\title{Move morphological descriptor terms to tax_epithet}
\usage{
move_morphological_descriptors(df)
}
\arguments{
\item{df}{A data frame processed by \code{\link[=move_uncertainty_descriptors]{move_uncertainty_descriptors()}}.}
}
\value{
The input data frame with updated \code{taxon_clean} and \code{tax_epithet}.
}
\description{
Detects terms like \code{centric}, \code{pennate}, \code{thecate}, \code{armored}, etc. and
moves them from \code{taxon_clean} to \code{tax_epithet}.
}
