% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{move_reproductive_structures}
\alias{move_reproductive_structures}
\title{Move reproductive and morphological structure terms to tax_epithet}
\usage{
move_reproductive_structures(df)
}
\arguments{
\item{df}{A data frame processed by \code{\link[=remove_dots]{remove_dots()}}.}
}
\value{
The input data frame with updated \code{taxon_clean} and \code{tax_epithet}.
}
\description{
Detects terms like \code{cysts}, \code{spores}, \code{colonial}, \code{benthic}, etc. and
moves them from \code{taxon_clean} to \code{tax_epithet}.
}
