% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{move_uncertainty_descriptors}
\alias{move_uncertainty_descriptors}
\title{Move uncertainty descriptor terms to tax_epithet}
\usage{
move_uncertainty_descriptors(df)
}
\arguments{
\item{df}{A data frame processed by \code{\link[=move_reproductive_structures]{move_reproductive_structures()}}.}
}
\value{
The input data frame with updated \code{taxon_clean}, \code{tax_epithet},
and \code{uncertain}.
}
\description{
Detects terms like \code{unidentified}, \code{undetermined}, \code{unclassified}, etc.
and moves them from \code{taxon_clean} to \code{tax_epithet}, also setting
\code{uncertain = TRUE}.
}
