% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{move_with_descriptors}
\alias{move_with_descriptors}
\title{Move 'with ...' clauses to tax_epithet}
\usage{
move_with_descriptors(df)
}
\arguments{
\item{df}{A data frame processed by \code{\link[=move_morphological_descriptors]{move_morphological_descriptors()}}.}
}
\value{
The input data frame with updated \code{taxon_clean} and \code{tax_epithet}.
}
\description{
Detects \verb{with ...} trailing clauses in \code{taxon_clean} and moves them to
\code{tax_epithet}.
}
