% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{normalize_infraspecific_ranks}
\alias{normalize_infraspecific_ranks}
\title{Normalize infraspecific rank notation}
\usage{
normalize_infraspecific_ranks(df)
}
\arguments{
\item{df}{A data frame processed by \code{\link[=process_epithet_entries]{process_epithet_entries()}}.}
}
\value{
The input data frame with normalized \code{taxon_clean}.
}
\description{
Ensures consistent formatting of \code{var.}, \code{f.}, \code{subsp.}, \code{ssp.}, \code{cv.}
in \code{taxon_clean} (space before, dot after, space after dot).
}
