% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{op-null-default}
\alias{op-null-default}
\alias{\%||\%}
\title{Null-coalescing operator}
\usage{
lhs \%||\% rhs
}
\arguments{
\item{lhs}{An object.}

\item{rhs}{Fallback value returned when \code{lhs} is \code{NULL}.}
}
\value{
\code{lhs} if not \code{NULL}, else \code{rhs}.
}
\description{
Returns \code{lhs} if not \code{NULL}, otherwise returns \code{rhs}.
}
\keyword{internal}
