% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{process_bracket_entries}
\alias{process_bracket_entries}
\title{Extract bracket annotations (\verb{-[...]})}
\usage{
process_bracket_entries(df)
}
\arguments{
\item{df}{A data frame processed by \code{\link[=process_sp_entries]{process_sp_entries()}}.}
}
\value{
The input data frame with updated \code{taxon_clean} and \code{tax_epithet}.
}
\description{
Moves bracket-style annotations of the form \verb{-[...]} from \code{taxon_clean}
to \code{tax_epithet}.
}
