% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{process_epithet_entries}
\alias{process_epithet_entries}
\title{Extract remaining parenthetical epithets}
\usage{
process_epithet_entries(df)
}
\arguments{
\item{df}{A data frame processed by \code{\link[=move_size_to_epithet]{move_size_to_epithet()}}.}
}
\value{
The input data frame with updated \code{taxon_clean} and \code{tax_epithet}.
}
\description{
Moves any remaining content in parentheses from \code{taxon_clean} to
\code{tax_epithet}.
}
