% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{process_sp_entries}
\alias{process_sp_entries}
\title{Extract sp./spp. designations}
\usage{
process_sp_entries(df)
}
\arguments{
\item{df}{A data frame processed by \code{\link[=process_incertae_entries]{process_incertae_entries()}}.}
}
\value{
The input data frame with updated \code{taxon_clean} and \code{tax_epithet}.
}
\description{
Moves \code{sp.}, \code{spp.}, \code{sp1}, etc. from \code{taxon_clean} to \code{tax_epithet}.
}
