% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{process_taxonomic_prefixes}
\alias{process_taxonomic_prefixes}
\title{Extract rank prefixes (O., C., F., P.)}
\usage{
process_taxonomic_prefixes(df)
}
\arguments{
\item{df}{A data frame processed by \code{\link[=normalize_characters]{normalize_characters()}}.}
}
\value{
The input data frame with updated \code{taxon_clean}, \code{tax_epithet},
and \code{uncertain}.
}
\description{
Detects and removes rank abbreviation prefixes such as \code{O.}, \code{C.}, \code{F.},
\code{P.} from \code{taxon_clean}, storing rank information in \code{tax_epithet}.
}
